bug with router in Profile
