+++ src/components/OrderDetails/OrderDetails.jsx
@@ -7,9 +7,10 @@ import Preloader from "../Preloader/Preloader";
 
 /* Соержимое модалки с деталями заказа, которые устанваливаются кликом при формировании заказа */
 const OrderDetails = () => {
+
   const currentOrderNumber = useSelector(store => store.orderDetails.currentOrderNumber);
-  const isLoading = useSelector(store => store.orderDetails.orederRequest); 
-  const isRequestSuccess = useSelector(store => store.orderDetails.orderSuccess);
+  /* const isLoading = useSelector(store => store.orderDetails.orederRequest); 
+  const isRequestSuccess = useSelector(store => store.orderDetails.orderSuccess); */
   const requestFailed = useSelector(store => store.orderDetails.orderfailed)
   
   {!currentOrderNumber && (
@@ -18,7 +19,7 @@ const OrderDetails = () => {
   if (currentOrderNumber) {
     return (
       <div className={`${styles.orderDetails} pt-30 pb-30`}>
-        <h3 className='styles.orderDetails__title text text_type_digits-large'>{ currentOrderNumber }</h3>
+        <h3 className={`${styles.orderDetails__title} text text_type_digits-large`}>{ currentOrderNumber }</h3>
         <p className='text text_type_main-medium pt-8 pb-15'>идентификатор заказа</p>
         <img className={styles.orderDetails__image} src={require('./images/order accpeted-popup-graphics.png')} alt='изображение готовности заказа' />
         <p className='text text_type_main-default pt-15 pb-2'>Ваш заказ начали готовить</p>
