+++ src/components/OrderDetails/orderDetails.module.css
@@ -8,6 +8,7 @@
 
 .orderDetails__title {
   padding-bottom: 10px;
+  text-shadow: 0px 0px 16px rgba(51, 51, 255, 0.25), 0px 0px 8px rgba(51, 51, 255, 0.25), 0px 4px 32px rgba(51, 51, 255, 0.5);
 }
 
 .orderDetails__image {
