+++ src/components/ProfileForm/profileForm.module.css
@@ -1,0 +1,4 @@
+.profileForm__buttons {
+  display: flex;
+  justify-content: flex-end;
+}
