+++ src/pages/ProfilePage/profilePage.module.css
@@ -35,8 +35,3 @@
   padding-left: 300px;
   margin: 0;
 }
-
-.profilePage__formButtons {
-  display: flex;
-  justify-content: flex-end;
-}
